backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
httpx
python-dotenv
hypothesis
pytest
pytest-asyncio
